Remove some unused confs

+++ install.sh
@@ -11,13 +11,6 @@ type nvim > /dev/null 2>&1 && rm -rf .config/nvim && ln -s ~/.vim .config/nvim
 type nvim > /dev/null 2>&1 && nvim +PlugInstall +qall || vim +PlugInstall +qall
 
 cd ~/.vim
-if [[ ! -f init.vim ]]; then
-    ln -s vimrc init.vim
-fi
+[[ ! -f init.vim ]] && ln -s vimrc init.vim
 
-# compile YouCompleteMe
-#cd ~/.vim/plugged/YouCompleteMe
-#type cargo > /dev/null 2>&1 && RUST="--rust-completer" || RUST=""
-#python3 install.py --clang-completer $RUST
-#cd -
-mkdir cache
+[[ ! -d cache ]] && mkdir cache
